serviccard-moved-to-customer
